fix(config): defer missing partial name; share schema isolation; cut complexity
Address review feedback on the config defer-validation PR.

- partial_encryption no longer eagerly crashes load_config on a missing `name`:
  from_dict now reads env.get("name", "") instead of env["name"], so a typo in
  this section can't KeyError every command (encrypt/decrypt/guard/pull/sync)
  that never reads it. validate() now also requires a non-empty name, surfacing
  the error only when a partial command consumes the section (cubic P2, #413).

- agent register now stores the normalized idle_timeout back onto the config
  (e.g. "10M" -> "10m") instead of discarding validate()'s return (greptile P2).

- schema: extract a shared _isolated_import() context manager used by both
  load() and get_schema_metadata_func(), removing the duplicated sys.path /
  sys.modules isolation. get_schema_metadata() is now invoked *inside* the
  isolation context, so a lazy sibling import in a user's metadata function
  resolves against its own service dir, not the restored caller environment.

- reduce CodeScene complexity flagged on the PR: split from_dict into per-section
  _build_*_config helpers (132 -> ~41 lines), extract _report_guardian_status
  from register, and route push/pull partial validation through one shared
  _validate_partial_config_or_exit helper.

- split test_config.py: move the deferred guardian/partial/EXAMPLE_CONFIG tests
  into test_config_deferred_validation.py so both files sit under the single-file
  size threshold.

Regression tests: missing-name defers instead of KeyError-ing; a lazy sibling
import inside get_schema_metadata() resolves to the right service.

Refs #413
